pyproject.toml

[project]
name = "treasury-forecasting-regime-nn"
version = "1.0.1"
description = "Regime-aware neural networks for forecasting the 10-Year U.S. Treasury yield"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "Dante Shoghanian"}]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "pyyaml>=6.0.1",
    "torch>=2.2",
    "pandas_datareader>=0.10",
    "fredapi>=0.5",
    "tqdm>=4.66"
]

[project.scripts]
treasury-train = "treasury_forecaster.training.train:main"
treasury-download = "treasury_forecaster.data.fred_loader:cli"
treasury-features = "treasury_forecaster.data.features:cli"
treasury-ingest = "treasury_forecaster.data.ingest_existing:cli"

[tool.setuptools.packages.find]
where = ["src"]
